ObjMD.AmzRestore: add content-md5 field
This new (optional) field is used to store the actual md5 of the
restored object, when it does not match the original content-md5. This
happens for multi-part objects, when the object is not restored with the
same number of parts; and we must in that case keep the original
content-md5 in metadata to ensure the reported ETag does not change.

Issue: ARSN-432