Extend the route types to support non-arsenal errors
Arsenal routes today are used in cloudserver, that calls these
callback functions using, most of the time, Arsenal errors. But
sometimes, especially when using modules, such as hdclient or
others, we might not return an Arsenal error. In this case, we
still try to access undefined properties, eventually leading to
strange responses to the client.

The changes are introducing support for traditional Error objects,
so that we can safely pass any error from a module used in an API
call.