ARSN-494 [tests] new test cases for ReplicationConfiguration
Add test cases to cover most of the validation of the
ReplicationConfiguration model, which did not have unit test coverage.

Also polished some error messages and the code style, as we now have a
more relaxed linter than when the code was originally written.

Note that the error types returned are kept and tested as they are
now, even if they may benefit from a more consistent handling, to
avoid any risk of regression due to a behavior change. We could
revisit this later, separately from this bugfix.