Ensure we detect properly bulk error when creating versions
We used to rely on a non-working detection algorithm that always
detected that the versionID operation failed, even if it was
the master version. As a result, the retry would recreate the
versionID a second time, and lead to more race conditions and
invalid states.

Now, we properly detect the index of the failing operation.
For versionID: we retry. For mAsters: we ignore, it will be
repaired.

Issue: ARSN-491