GCP client migration
As we can no longer use the Service.defineService
previously used in aws-sdk v2, we have to manually
this commit aims to implement the GCP client
using the new aws-sdk v3 architecture
and migrate the existing code to use it.
Please note that now both the setup and signing
mechanisms had to be re-implemented in the GcpService
file.
Issue: ARSN-514