ARSN-541 [impr] storage helper OptionalDecompressStream
Add a helper stream that detects whether the input is compressed or
not, based on a magic number in the very first bytes, and if a match
is found, transparently forward the decompressed payload to the
output, otherwise, just forward the original input payload. It works
fine as long as we control the format of the input to ensure its
initial bytes never collide with an existing supported magic
number. In particular, a JSON payload is safe in this respect.

The intended use case is to be able to compress Metadata backup chunks
and decompress them transparently. Since we don't have to store
metadata about compression, we can re-compress existing backup chunks
by simply putting them again on the RING in a compressed form.

Zstandard is currently the only compression scheme supported.